Доработка апи пользователей

+++ src/api/v1/users/routes.py
@@ -1,11 +1,11 @@
 from flask import jsonify
 from flask_jwt_extended import set_access_cookies
 from flask_openapi3 import APIBlueprint, Tag
-from flask_security import permissions_accepted
+from flask_security import permissions_accepted, current_user
 
 from api.v1.users.services import UserService
 from api.v1.users.schemas import CreateUserS, BaseUserS, ReadUserS, LoginS, LoggedInS, UserPath
-from errors import WasNotFoundError
+from errors import WasNotFoundError, ForbiddenError
 from global_schemas import PaginationQS, ErrorS
 from global_schemas import security_schemas
 from security import jwt_required
@@ -22,28 +22,26 @@ auth_router = APIBlueprint(
 )
 
 
-@users_router.post('/', responses={201: ReadUserS, 409: ErrorS})
-@jwt_required()
-@permissions_accepted('user-write')
-def add_user(body: CreateUserS):
-    created_user = UserService.add(body)
-    return jsonify(created_user.model_dump()), 201
+# @users_router.post('/', responses={201: ReadUserS, 409: ErrorS})
+# @jwt_required()
+# def add_user(body: CreateUserS):
+#     created_user = UserService.add(body)
+#     return jsonify(created_user.model_dump()), 201
 
 
-@users_router.get('/', responses={200: ReadUserS, 400: ErrorS})
-@jwt_required()
-@permissions_accepted('user-read')
-def get_users(query: PaginationQS):
-    users = UserService.get_many(query.limit, query.page)
-    return jsonify([user.model_dump() for user in users]), 200
+# @users_router.get('/', responses={200: ReadUserS, 400: ErrorS})
+# @jwt_required()
+# def get_users(query: PaginationQS):
+#     return jsonify([]), 200
+#     # users = UserService.get_many(query.limit, query.page)
+#     # return jsonify([user.model_dump() for user in users]), 200
 
 
 @users_router.get('/<int:user_id>/', responses={200: ReadUserS, 404: ErrorS})
 @jwt_required()
-@permissions_accepted('user-read')
 def get_user_by_id(path: UserPath):
     user = UserService.get_one_by_id_or_none(path.user_id)
-    if user is None:
+    if not user:
         raise WasNotFoundError(f'User with id {path.user_id}')
 
     return jsonify(user.model_dump()), 200
@@ -51,8 +49,9 @@ def get_user_by_id(path: UserPath):
 
 @users_router.put('/<int:user_id>/', responses={200: ReadUserS, 404: ErrorS, 409: ErrorS})
 @jwt_required()
-@permissions_accepted('user-write')
 def update_user_by_id(path: UserPath, body: BaseUserS):
+    if current_user.id != path.user_id:
+        raise ForbiddenError()
     updated_user_schema = UserService.update_by_id(path.user_id, body)
     return jsonify(updated_user_schema.model_dump()), 200
 
@@ -61,6 +60,8 @@ def update_user_by_id(path: UserPath, body: BaseUserS):
 @jwt_required()
 @permissions_accepted('user-write')
 def delete_user_by_id(path: UserPath):
+    if current_user.id != path.user_id:
+        raise ForbiddenError()
     UserService.delete_by_id(path.user_id)
     return jsonify(), 204
